Move to hcloud v2 client (mainly int64 migrations

+++ internal/context_menu/context_menu.go
@@ -15,5 +15,5 @@ type ContextMenu struct {
 	Items        []ContextMenuItem
 	SelectedItem int
 	ResourceType resource.ResourceType
-	ResourceID   int
+	ResourceID   int64
 }
